fix get_storage_region
